Broke out requst interface into own file.
